Run preprocessingMinfiEwasWater.R as an external script from inst/script/.R Run preprocessingMinfiEwasWater as an external pipeline script

Wrapper function that executes the preprocessingMinfiEwasWater.R script shipped in inst/scripts using system calls.

preprocessingMinfiEwasWater(
  phenoFile = "data/preprocessingMinfiEwasWater/pheno.csv",
  idatFolder = "data/preprocessingMinfiEwasWater/idats",
  outputLogs = "logs",
  nSamples = NA,
  SampleID = "Sample_Name",
  arrayType = "IlluminaHumanMethylationEPICv2",
  annotationVersion = "20a1.hg38",
  scriptLabel = "preprocessingMinfiEwasWater",
  baseDataFolder = "rData",
  sepType = "",
  tiffWidth = 2000,
  tiffHeight = 1000,
  tiffRes = 150,
  qcCutoff = 10.5,
  detPtype = "m+u",
  detPThreshold = 0.05,
  funnormSeed = 123,
  normMethods = "adjustedfunnorm",
  sexColumn = "Sex",
  pvalThreshold = 0.01,
  chrToRemove = "chrX,chrY",
  snpsToRemove = "SBE,CpG",
  mafThreshold = 0.1,
  crossReactivePath = "data/preprocessingMinfiEwasWater/12864_2024_10027_MOESM8_ESM.csv",
  plotGroupVar = "Sex",
  lcRef = "salivaEPIC",
  phenoOrder = "Sample_Name;Timepoint;Sex;PredSex;Basename;Sentrix_ID;Sentrix_Position",
  lcPhenoDir = "data/preprocessingMinfiEwasWater"
)

Arguments

phenoFile

Character. Path to phenotype CSV file.

idatFolder

Character. Path to IDAT files directory.

outputLogs

Character. Directory for log files.

nSamples

Integer or NA. Number of samples to subset for testing.

SampleID

Character. Sample identifier column name.

arrayType

Character. Illumina array type.

annotationVersion

Character. Annotation version.

scriptLabel

Character. Label used in output naming.

baseDataFolder

Character. Base directory for RData outputs.

sepType

Character. Field separator for phenotype file.

tiffWidth

Integer. Width of TIFF plots in pixels.

tiffHeight

Integer. Height of TIFF plots in pixels.

tiffRes

Integer. Resolution (DPI) for TIFF plots.

qcCutoff

Numeric. Quality-control cutoff threshold.

detPtype

Character. Detection p-value calculation type.

detPThreshold

Numeric. Detection p-value threshold.

funnormSeed

Integer. Random seed for normalization.

normMethods

Character. Normalization method(s).

sexColumn

Character. Sex column name in phenotype data.

pvalThreshold

Numeric. Probe-level p-value threshold.

chrToRemove

Character. Chromosomes to remove.

snpsToRemove

Character. SNP probe types to remove.

mafThreshold

Numeric. Minor allele frequency threshold.

crossReactivePath

Character. Path to cross-reactive probe file.

plotGroupVar

Character. Variable used for grouping plots.

lcRef

Character. Reference panel for cell composition.

phenoOrder

Character. Semicolon-separated phenotype column order.

lcPhenoDir

Character. Output directory for cell composition phenotype.

Value

Invisibly returns NULL. This function is called for its side effects, performing Illumina EPICv2 preprocessing, quality control, normalisation, probe filtering, cell composition estimation, and writing plots, logs, and RData objects to disk.

Examples

tmp <- tempdir()
stopifnot(dir.exists(tmp))

# \donttest{
preprocessingMinfiEwasWater(
  phenoFile = "data/preprocessingMinfiEwasWater/pheno.csv",
  idatFolder = "data/preprocessingMinfiEwasWater/idats",
  outputLogs = "logs",
  nSamples = NA,
  SampleID = "Sample_Name",
  arrayType = "IlluminaHumanMethylationEPICv2",
  annotationVersion = "20a1.hg38",
  scriptLabel = "preprocessingMinfiEwasWater",
  baseDataFolder = "rData",
  sepType = "",
  tiffWidth = 2000,
  tiffHeight = 1000,
  tiffRes = 150,
  qcCutoff = 10.5,
  detPtype = "m+u",
  detPThreshold = 0.05,
  funnormSeed = 123,
  normMethods = "adjustedfunnorm",
  sexColumn = "Sex",
  pvalThreshold = 0.01,
  chrToRemove = "chrX,chrY",
  snpsToRemove = "SBE,CpG",
  mafThreshold = 0.1,
  crossReactivePath = "data/preprocessingMinfiEwasWater/12864_2024_10027_MOESM8_ESM.csv",
  plotGroupVar = "Sex",
  lcRef = "salivaEPIC",
  phenoOrder = "Sample_Name;Timepoint;Sex;PredSex;Basename;Sentrix_ID;Sentrix_Position",
  lcPhenoDir = "data/preprocessingMinfiEwasWater"
)
#> Running preprocessingMinfiEwasWater:
#> Rscript "C:/Users/ruizpint/AppData/Local/Programs/R/R-4.4.1/library/dnaEPICO/scripts/preprocessingMinfiEwasWater.R" --phenoFile "data/preprocessingMinfiEwasWater/pheno.csv" --sepType "" --idatFolder "data/preprocessingMinfiEwasWater/idats" --outputLogs "logs" --nSamples NA --SampleID "Sample_Name" --arrayType "IlluminaHumanMethylationEPICv2" --annotationVersion "20a1.hg38" --scriptLabel "preprocessingMinfiEwasWater" --baseDataFolder "rData" --tiffWidth 2000 --tiffHeight 1000 --tiffRes 150 --qcCutoff 10.5 --detPtype "m+u" --detPThreshold 0.05 --funnormSeed 123 --normMethods "adjustedfunnorm" --sexColumn "Sex" --pvalThreshold 0.01 --chrToRemove "chrX,chrY" --snpsToRemove "SBE,CpG" --mafThreshold 0.1 --crossReactivePath "data/preprocessingMinfiEwasWater/12864_2024_10027_MOESM8_ESM.csv" --plotGroupVar "Sex" --lcRef "salivaEPIC" --phenoOrder "Sample_Name;Timepoint;Sex;PredSex;Basename;Sentrix_ID;Sentrix_Position" --lcPhenoDir "data/preprocessingMinfiEwasWater"
#> Warning: 'Rscript "C:/Users/ruizpint/AppData/Local/Programs/R/R-4.4.1/library/dnaEPICO/scripts/preprocessingMinfiEwasWater.R" --phenoFile "data/preprocessingMinfiEwasWater/pheno.csv" --sepType "" --idatFolder "data/preprocessingMinfiEwasWater/idats" --outputLogs "logs" --nSamples NA --SampleID "Sample_Name" --arrayType "IlluminaHumanMethylationEPICv2" --annotationVersion "20a1.hg38" --scriptLabel "preprocessingMinfiEwasWater" --baseDataFolder "rData" --tiffWidth 2000 --tiffHeight 1000 --tiffRes 150 --qcCutoff 10.5 --detPtype "m+u" --detPThreshold 0.05 --funnormSeed 123 --normMethods "adjustedfunnorm" --sexColumn "Sex" --pvalThreshold 0.01 --chrToRemove "chrX,chrY" --snpsToRemove "SBE,CpG" --mafThreshold 0.1 --crossReactivePath "data/preprocessingMinfiEwasWater/12864_2024_10027_MOESM8_ESM.csv" --plotGroupVar "Sex" --lcRef "salivaEPIC" --phenoOrder "Sample_Name;Timepoint;Sex;PredSex;Basename;Sentrix_ID;Sentrix_Position" --lcPhenoDir "data/preprocessingMinfiEwasWater"' execution failed with error code 1
# }