Run svaEnmix as an external pipeline script

Wrapper function that executes the svaEnmix.R script located in inst/scripts using a system call.

svaEnmix(
  phenoFile = "data/preprocessingMinfiEwasWater/phenoLC.csv",
  rgsetData = "rData/preprocessingMinfiEwasWater/objects/RGSet.RData",
  sepType = "",
  outputLogs = "logs",
  nSamples = NA,
  SampleID = "Sample_Name",
  arrayType = "IlluminaHumanMethylationEPICv2",
  annotationVersion = "20a1.hg38",
  SentrixIDColumn = "Sentrix_ID",
  SentrixPositionColumn = "Sentrix_Position",
  ctrlSvaPercVar = 0.9,
  ctrlSvaFlag = 1,
  scriptLabel = "svaEnmix",
  tiffWidth = 2000,
  tiffHeight = 1000,
  tiffRes = 150,
  figureBaseDir = "figures",
  dataBaseDir = "data"
)

Arguments

phenoFile

Character. Path to phenotype file with cell composition data.

rgsetData

Character. Path to RGSet RData file.

sepType

Character. Field separator for phenotype file.

outputLogs

Character. Directory for log files.

nSamples

Integer or NA. Number of samples to subset for testing.

SampleID

Character. Sample identifier column name.

arrayType

Character. Illumina array type.

annotationVersion

Character. Annotation version.

SentrixIDColumn

Character. Sentrix ID column name.

SentrixPositionColumn

Character. Sentrix position column name.

ctrlSvaPercVar

Numeric. Proportion of variance explained by control probes.

ctrlSvaFlag

Integer. Flag indicating use of control probes.

scriptLabel

Character. Label used in output naming.

tiffWidth

Integer. Width of TIFF plots in pixels.

tiffHeight

Integer. Height of TIFF plots in pixels.

tiffRes

Integer. Resolution (DPI) for TIFF plots.

figureBaseDir

Character. Base directory for Figures outputs.

dataBaseDir

Character. Base directory for Data outputs.

Value

Invisibly returns NULL. This function is called for its side effects, executing the external svaEnmix.R script and writing results, figures, and logs to disk.

Examples

tmp <- tempdir()
stopifnot(dir.exists(tmp))

# \donttest{
svaEnmix(
  phenoFile = "data/preprocessingMinfiEwasWater/phenoLC.csv",
  rgsetData = "rData/preprocessingMinfiEwasWater/objects/RGSet.RData",
  sepType = "",
  outputLogs = "logs",
  nSamples = 5,
  SampleID = "Sample_Name",
  arrayType = "IlluminaHumanMethylationEPICv2",
  annotationVersion = "20a1.hg38",
  SentrixIDColumn = "Sentrix_ID",
  SentrixPositionColumn = "Sentrix_Position",
  ctrlSvaPercVar = 0.90,
  ctrlSvaFlag = 1,
  scriptLabel = "svaEnmix",
  tiffWidth = 2000,
  tiffHeight = 1000,
  tiffRes = 150,
  figureBaseDir = "figures",
  dataBaseDir = "data"
)
#> Running svaEnmix:
#> Rscript "C:/Users/ruizpint/OneDrive - Queensland University of Technology/QUT/FinalResults/DNAm/dnaEPICO/inst/scripts/svaEnmix.R" --phenoFile "data/preprocessingMinfiEwasWater/phenoLC.csv" --rgsetData "rData/preprocessingMinfiEwasWater/objects/RGSet.RData" --sepType "" --outputLogs "logs" --nSamples 5 --SampleID "Sample_Name" --arrayType "IlluminaHumanMethylationEPICv2" --annotationVersion "20a1.hg38" --SentrixIDColumn "Sentrix_ID" --SentrixPositionColumn "Sentrix_Position" --ctrlSvaPercVar 0.9 --ctrlSvaFlag 1 --scriptLabel "svaEnmix" --tiffWidth 2000 --tiffHeight 1000 --tiffRes 150 --figureBaseDir "figures" --dataBaseDir "data"
#> Warning: 'Rscript "C:/Users/ruizpint/OneDrive - Queensland University of Technology/QUT/FinalResults/DNAm/dnaEPICO/inst/scripts/svaEnmix.R" --phenoFile "data/preprocessingMinfiEwasWater/phenoLC.csv" --rgsetData "rData/preprocessingMinfiEwasWater/objects/RGSet.RData" --sepType "" --outputLogs "logs" --nSamples 5 --SampleID "Sample_Name" --arrayType "IlluminaHumanMethylationEPICv2" --annotationVersion "20a1.hg38" --SentrixIDColumn "Sentrix_ID" --SentrixPositionColumn "Sentrix_Position" --ctrlSvaPercVar 0.9 --ctrlSvaFlag 1 --scriptLabel "svaEnmix" --tiffWidth 2000 --tiffHeight 1000 --tiffRes 150 --figureBaseDir "figures" --dataBaseDir "data"' execution failed with error code 1
# }